added classes
added the functionality to grade a student at a specific subject (given the student_id and the subject_id).
the grades list stores the grades for all subjects.
It will look like this for each student:
grades[student_id] = [
[subject_id1, grade1, grade2],
[subject_id2, grade1],
.........,
[subject_idN, grade1, grade2, grade3]
]

Each time we check if the student already has at least one grade for the given subject_id, if so we just push the grade into the given subject_id list, if not: we create it.